+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 
 * Use Reline for interactive mode of `hexapdf inspect` if available
+* [HexaPDF::DictionaryFields::Field::new] to use keyword arguments
 
 
 ## 0.10.0 - 2019-10-02